// src/common/models/device-registration.class.ts
export class DeviceRegistration{
    deviceId: string;
    pushId: any;
    awsNotificationToken: string;
    model: string;
    devicePlatform: string;
    osVersion: any;
    type:string
    status: string;
    hardwareId: string;
    numberOfSimSlot: string;
    userAgent: string;
    IMEI: string;
    os: any;
    deviceType: string;
}